chore: 동화 도메인 개발을 위한 패키지 구성 변경
