--- src/main/java/hanium/server/i_luv_book/fairytale/PageOptionInfo.java
@@ -1,4 +1,0 @@
-package hanium.server.i_luv_book.fairytale;
-
-public class PageOptionInfo {
-}
