+++ src/main/java/hanium/server/i_luv_book/fairytale/domain/Keyword.java
@@ -1,0 +1,4 @@
+package hanium.server.i_luv_book.fairytale.domain;
+
+public class Keyword {
+}
