+++ src/main/java/hanium/server/i_luv_book/security/authorize/manager/CustomAuthorizationManager.java
@@ -5,7 +5,6 @@ import hanium.server.i_luv_book.security.authorize.service.AuthorizationService;
 import jakarta.annotation.PostConstruct;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.core.log.LogMessage;
 import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
 import org.springframework.security.authorization.AuthorityAuthorizationManager;
 import org.springframework.security.authorization.AuthorizationDecision;
