+++ src/main/java/hanium/server/i_luv_book/security/exception/entrypoint/CustomAuthenticationEntryPoint.java
@@ -6,7 +6,6 @@ import hanium.server.i_luv_book.security.exception.code.SecurityExceptionCode;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
